+++ cmd/falco/main.go
@@ -6,11 +6,12 @@ import (
 	"os"
 	"strings"
 
-	"path/filepath"
+	"encoding/json"
 
 	"github.com/fatih/color"
 	"github.com/kyokomi/emoji"
 	"github.com/mattn/go-colorable"
+	"github.com/pkg/errors"
 )
 
 var version string = ""
@@ -23,6 +24,13 @@ var (
 	green   = color.New(color.FgGreen)
 	cyan    = color.New(color.FgCyan)
 	magenta = color.New(color.FgMagenta)
+
+	ErrExit = errors.New("exit")
+)
+
+const (
+	subcommandLint      = "lint"
+	subcommandTerraform = "terraform"
 )
 
 type multiStringFlags []string
@@ -44,6 +52,8 @@ type Config struct {
 	VV           bool
 	Version      bool
 	Remote       bool
+	Stats        bool
+	Json         bool
 }
 
 func write(c *color.Color, format string, args ...interface{}) {
@@ -60,7 +70,11 @@ func printUsage() {
   falco: Fastly VCL parser / linter
 =======================================
 Usage:
-    falco [main vcl file]
+    falco [subcommand] [main vcl file]
+
+Subcommands:
+    terraform : Run lint from terraform planned JSON
+    lint      : Run lint (default)
 
 Flags:
     -I, --include_path : Add include path
@@ -70,9 +84,18 @@ Flags:
     -V, --version      : Display build version
     -v                 : Verbose warning lint result
     -vv                : Varbose all lint result
+    -json              : Output statistics as JSON
+    -stats             : Analyze VCL statistics
 
-Example:
+Simple Linting example:
     falco -I . -vv /path/to/vcl/main.vcl
+
+Get statistics example:
+    falco -I . -stats /path/to/vcl/main.vcl
+
+Linting with terraform:
+    terraform plan -out planned.out
+    terraform show -json planned.out | falco -vv terraform
 `
 
 	fmt.Println(strings.TrimLeft(usage, "\n"))
@@ -94,8 +117,12 @@ func main() {
 	fs.BoolVar(&c.Version, "version", false, "Print Version")
 	fs.BoolVar(&c.Remote, "r", false, "Use Remote")
 	fs.BoolVar(&c.Remote, "remote", false, "Use Remote")
+	fs.BoolVar(&c.Stats, "stats", false, "Analyze VCL statistics")
+	fs.BoolVar(&c.Json, "json", false, "Output statistics as JSON")
+	fs.Usage = printUsage
 
-	if err := fs.Parse(os.Args[1:]); err != nil {
+	var err error
+	if err = fs.Parse(os.Args[1:]); err != nil {
 		if err == flag.ErrHelp {
 			printUsage()
 		}
@@ -109,36 +136,60 @@ func main() {
 		os.Exit(1)
 	}
 
-	mainVcl := fs.Arg(0)
-	if mainVcl == "" {
-		printUsage()
-	} else if _, err := os.Stat(mainVcl); err != nil {
-		if err == os.ErrNotExist {
-			writeln(white, "Input file %s is not found", mainVcl)
-		} else {
-			writeln(red, "Unexpected stat error: %s", err.Error())
-		}
-		os.Exit(1)
+	// falco could lint multiple services so resolver should be a slice
+	var resolvers []Resolver
+	switch fs.Arg(0) {
+	case subcommandTerraform:
+		resolvers, err = NewStdinResolvers()
+	case subcommandLint:
+		// "lint" command provides single file of service, then resolvers size is always 1
+		resolvers, err = NewFileResolvers(fs.Arg(1), c)
+	default:
+		// "lint" command provides single file of service, then resolvers size is always 1
+		resolvers, err = NewFileResolvers(fs.Arg(0), c)
 	}
 
-	vcl, err := filepath.Abs(mainVcl)
-	if err != nil {
-		writeln(red, "Failed to get absolute path: %s", err.Error())
-		os.Exit(1)
-	}
-
-	runner, err := NewRunner(vcl, c)
 	if err != nil {
 		writeln(red, err.Error())
 		os.Exit(1)
 	}
 
+	var shouldExit bool
+	for _, v := range resolvers {
+		if name := v.Name(); name != "" {
+			writeln(white, `Lint service of "%s"`, name)
+			writeln(white, strings.Repeat("=", 18+len(name)))
+		}
+
+		runner, err := NewRunner(v, c)
+		if err != nil {
+			writeln(red, err.Error())
+			os.Exit(1)
+		}
+
+		var exitErr error
+		if c.Stats {
+			exitErr = runStats(runner, c.Json)
+		} else {
+			exitErr = runLint(runner)
+		}
+		if exitErr == ErrExit {
+			shouldExit = true
+		}
+	}
+
+	if shouldExit {
+		os.Exit(1)
+	}
+}
+
+func runLint(runner *Runner) error {
 	result, err := runner.Run()
 	if err != nil {
 		if err != ErrParser {
 			writeln(red, err.Error())
 		}
-		os.Exit(1)
+		return ErrExit
 	}
 
 	write(red, ":fire:%d errors, ", result.Errors)
@@ -168,11 +219,57 @@ func main() {
 		if len(runner.transformers) > 0 {
 			writeln(white, "Program aborted. Please fix lint errors before transforming.")
 		}
-		os.Exit(1)
+		return ErrExit
 	}
 
 	if err := runner.Transform(result.Vcls); err != nil {
 		writeln(red, err.Error())
-		os.Exit(1)
+		return ErrExit
 	}
+	return nil
+}
+
+func runStats(runner *Runner, printJson bool) error {
+	stats, err := runner.Stats()
+	if err != nil {
+		if err != ErrParser {
+			writeln(red, err.Error())
+		}
+		return ErrExit
+	}
+
+	if printJson {
+		enc := json.NewEncoder(os.Stdout)
+		enc.SetIndent("", "  ")
+		if err := enc.Encode(stats); err != nil {
+			writeln(red, err.Error())
+			os.Exit(1)
+		}
+		return ErrExit
+	}
+
+	printStats(strings.Repeat("=", 80))
+	printStats("| %-76s |", "falco VCL statistics ")
+	printStats(strings.Repeat("=", 80))
+	printStats("| %-22s | %51s |", "Main VCL File", stats.Main)
+	printStats(strings.Repeat("=", 80))
+	printStats("| %-22s | %51d |", "Included Module Files", stats.Files-1)
+	printStats(strings.Repeat("-", 80))
+	printStats("| %-22s | %51d |", "Total VCL Lines", stats.Lines)
+	printStats(strings.Repeat("-", 80))
+	printStats("| %-22s | %51d |", "Subroutines", stats.Subroutines)
+	printStats(strings.Repeat("-", 80))
+	printStats("| %-22s | %51d |", "Backends", stats.Backends)
+	printStats(strings.Repeat("-", 80))
+	printStats("| %-22s | %51d |", "Tables", stats.Tables)
+	printStats(strings.Repeat("-", 80))
+	printStats("| %-22s | %51d |", "Access Control Lists", stats.Acls)
+	printStats(strings.Repeat("-", 80))
+	printStats("| %-22s | %51d |", "Directors", stats.Directors)
+	printStats(strings.Repeat("-", 80))
+	return nil
+}
+
+func printStats(format string, args ...interface{}) {
+	fmt.Fprintf(os.Stdout, format+"\n", args...)
 }